+++ extras/test/cores/avr/Stream.h
@@ -33,6 +33,8 @@ struct Stream : Print {
     return result;
   }
 
+  void setTimeout(unsigned long) {}
+
  protected:
   int timedRead() {
     return read();