+++ extras/test/cores/esp32/Stream.h
@@ -39,6 +39,8 @@ struct Stream : Print {
     return result;
   }
 
+  void setTimeout(unsigned long) {}
+
  protected:
   int timedRead() {
     return read();