+++ extras/test/cores/esp8266/Stream.h
@@ -36,6 +36,8 @@ struct Stream : Print {
     return result;
   }
 
+  void setTimeout(unsigned long) {}
+
  protected:
   int timedRead() {
     return read();