+++ extras/test/cores/samd/Stream.h
@@ -35,6 +35,8 @@ struct Stream : Print {
     return result;
   }
 
+  void setTimeout(unsigned long) {}
+
  protected:
   int timedRead() {
     return read();