+++ src/StreamUtils/Clients/ClientProxy.hpp
@@ -18,7 +18,7 @@ class ClientProxy : public Client {
                        ConnectPolicy connection)
       : _target(upstream),
         _reader(reader),
-        _writer(writer),
+        _writer(Polyfills::move(writer)),
         _connection(connection) {}
 
   ClientProxy(const ClientProxy &other)
@@ -94,7 +94,7 @@ class ClientProxy : public Client {
     _writer.flush(_target);
   }
 
- private:
+ protected:
   Client &_target;
   ReadPolicy _reader;
   WritePolicy _writer;